After quite a bit of struggle, found the right magic to make our model for events/entries work
Our event/entry model is really not quite the Rails Way, so in the end,
while some of the magic does work, the conventions must be bent.
There's a lot of cleanup still to be done here in how things are
displayed, but they do work.

Note: tests will probably fail, as they have not been updated. Bad
TDD-developer, no biscuit.

Former-commit-id: 20dc94361a0bfbebb95912a472db92b5e11123ea