Begins of role/permissions support
Events and Lost and Found are now governed by permissions bits which
are associated with roles, which in turn are associated with users.
That is, individual users do not have permissions-bits; they have one
or more assigned roles.

There are, however, helper functions on the User model that allow you
to treat the user as the one being permitted or denied something, and
these are what should actually be used when permitting or denying a
function.